Separated mock classes
