+++ packages/acter_notifify/test/mock_data/mock_notification_item.dart
@@ -1,0 +1,4 @@
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockNotificationItem extends Mock implements NotificationItem {}
