+++ packages/acter_notifify/test/mock_data/mock_notification_parent.dart
@@ -1,0 +1,5 @@
+
+import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockNotificationParent extends Mock implements NotificationItemParent {}
